+++ jest.config.js
@@ -1,0 +1,14 @@
+module.exports = {
+  testEnvironment: 'node',
+  setupFilesAfterEnv: ['./jest.setup.js'],
+  testMatch: ['**/test/**/*.test.js'],
+  collectCoverageFrom: ['src/**/*.js', '!src/**/*.test.js'],
+  coverageThreshold: {
+    global: {
+      branches: 70,
+      functions: 70,
+      lines: 70,
+      statements: 70,
+    },
+  },
+};